feat: Add logo image path and OpenCV library to build (portable-main)
This commit adds a path for the application logo image to the AppContext class and integrates the OpenCV library into the build process.  The logo path allows easy access for application-wide use. The OpenCV library is copied to the build directory to ensure proper functionality.  The version of the jar file is also updated.